pyproject.toml
[project]
name = "agentic-graysky"
version = "0.1.0"
description = "An API designed for AI agents to interact with the Graysky website content"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.103.0",
    "uvicorn>=0.23.2",
    "pydantic>=2.3.0",
    "sqlalchemy>=2.0.20",
    "python-multipart>=0.0.6",
    "markdown>=3.4.4",
    "python-frontmatter>=1.0.0",
    "jinja2>=3.1.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
